+++ CHANGELOG.md
@@ -17,6 +17,21 @@ Changelog structure reference:
 
 ## [Unreleased]
 
+## [14.3.0]
+
+### Added
+
+- Project
+  - Extra compiler arguments field for zig test/run tasks
+
+### Fixed
+
+- Debugging
+  - The debugger no longer freezes the IDE while zig is compiling
+
+- Project
+  - Exe args for zig run not visible in the GUI
+
 ## [14.2.0]
 
 ### Added